third_party/blink/web_tests/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-option-hover-styles-expected.txt

This is a testharness.js-based test.
[FAIL] Hover styles should be present for appearance:base-select options.
  assert_equals: Option background-color while hovering. expected "oklab(0 0 0 / 0.1)" but got "lab(0 0 0 / 0.1)"
Harness: the test ran to completion.
